chore(editor): Update PostCreateInput and PostUpdateInput types
